backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
python-dotenv==1.2.1
httpx==0.28.1
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3